+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/ApplicationPasswordQueryExecutionFixtureWebserverRequestTest.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace PHPUnitForGatoGraphQL\GatoGraphQL\Integration;
 
+use GuzzleHttp\RequestOptions;
+
 class ApplicationPasswordQueryExecutionFixtureWebserverRequestTest extends AbstractApplicationPasswordQueryExecutionFixtureWebserverRequestTestCase
 {
     use ApplicationPasswordQueryExecutionFixtureWebserverRequestTestTrait;
@@ -11,5 +13,26 @@ class ApplicationPasswordQueryExecutionFixtureWebserverRequestTest extends Abstr
     protected static function getEndpoint(): string
     {
         return 'graphql/';
+    }
+
+    /**
+     * Make the authentication fail
+     *
+     * @param array<string,mixed> $options
+     * @return array<string,mixed>
+     */
+    protected function customizeRequestOptions(array $options): array
+    {
+        $options = parent::customizeRequestOptions($options);
+
+        $dataName = $this->getDataName();
+        if (str_starts_with($dataName, 'error/')) {
+            $options[RequestOptions::HEADERS]['Authorization'] = sprintf(
+                'Basic %s',
+                base64_encode('admin:___fail___')
+            );
+        }
+
+        return $options;
     }
 }