+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Use `isGutenbergEditorEnabled` in predefined persisted queries (#2802)
 - Added mutations to assign custom tags/categories to custom posts (#2803)
 - Added Settings option to enable/disable logs (#2813)
+- Application password failed authentication: Show error in GraphQL response (#2817)
 
 ### Fixed